Tenants for squads

+++ src/controllers/squads/EditSquadPost.php
@@ -1,9 +1,13 @@
 <?php
 
 $db = app()->db;
+$tenant = app()->tenant;
 
-$getKey = $db->prepare("SELECT SquadKey FROM squads WHERE SquadID = ?");
-$getKey->execute([$id]);
+$getKey = $db->prepare("SELECT SquadKey FROM squads WHERE SquadID = ? AND Tenant = ?");
+$getKey->execute([
+  $id,
+  $tenant->getId()
+]);
 $squadDeleteKey = $getKey->fetchColumn();
 
 if (mb_strlen(trim($_POST['squadName'])) == 0) {
@@ -16,8 +20,11 @@ if ($_POST['squadFee'] < 0) {
 if ($_POST['squadDeleteDanger'] ==  $squadDeleteKey) {
   // Delete the squad
   try {
-    $delete = $db->prepare("DELETE FROM squads WHERE SquadID = ?");
-    $delete->execute([$id]);
+    $delete = $db->prepare("DELETE FROM squads WHERE SquadID = ? AND Tenant = ?");
+    $delete->execute([
+      $id,
+      $tenant->getId()
+    ]);
   } catch (Exception $e) {
     $_SESSION['UpdateDatabaseError'] = true;
   }
@@ -31,14 +38,15 @@ if ($_POST['squadDeleteDanger'] ==  $squadDeleteKey) {
 } else {
   // Update the squad
   try {
-    $update = $db->prepare("UPDATE squads SET SquadName = ?, SquadFee = ?, SquadCoach = ?, SquadTimetable = ?, SquadCoC = ? WHERE SquadID = ?");
+    $update = $db->prepare("UPDATE squads SET SquadName = ?, SquadFee = ?, SquadCoach = ?, SquadTimetable = ?, SquadCoC = ? WHERE SquadID = ? AND Tenant = ?");
     $update->execute([
       trim($_POST['squadName']),
       $_POST['squadFee'],
       trim($_POST['squadCoach']),
       trim($_POST['squadTimetable']),
       trim($_POST['squadCoC']),
-      $id
+      $id,
+      $tenant->getId()
     ]);
     $_SESSION['UpdateSuccess'] = true;
   } catch (Exception $e) {
